+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Enable `react/jsx-curly-brace-presence` with `'never'`
+- fix: Relax `@typescript-eslint/no-empty-interface` to `warn` level
+- fix: Relax `destructuring/in-params` to `warn` level
 
 ## 8.3.0